Added new annex.tune.noobjectdir setting.
Note that it only affects git repos, it does not affect special remotes
like directory, rsync, and even gcrypt that use keyPath themselves.

This seems to work pretty well, at first glance. I don't know for sure if
there's perhaps code somewhere outside of Locations.hs that assumes the
objectdir exists, except for the freeze/thaw code. Hopefully not;
I have always tried to keep all that code in one place.

XXX However, getKeysPresent is broken by this change!
See https://github.com/datalad/datalad/issues/32

---

Since the freeze/thaw looks at the parent directory of the object file, it
acts on the deepest level of hash directory in this configuration. This
works ok, when there's only one process adding/removing content at a time.
However, there's a race if two processes are both adding/removing content:

1. process A thaws directory
2. process B thaws directory (noop)
3. process A makes change
4. process A freezes directory
5. process B tries to make change in frozen directory, so fails

This race already perhaps existed, but only when both processes were
acting on the same key. With noobjectdir, processes can be acting on different
keys that share the same hash directory. If objecthash1 is also used, there
are only 1024 hash directories, so a conflict becomes likely.

Locking seems the best way to deal with this, but I have not put it in yet.